move attachFxHandlers to fx.js
